+++ Changelog.md
@@ -23,6 +23,7 @@ Breaking Changes:
  * Type System: Explicit conversions from literals to integer type is as strict as implicit conversions.
  * Type System: Support ``address(...).codehash`` to retrieve the codehash of an account.
  * Type System: Unary negation can only be used on signed integers, not on unsigned integers.
+ * View Pure Checker: Mark ``chainid`` as view.
 
 Language Features:
  * Super constructors can now be called using the member notation e.g. ``M.C(123)``.